Unquote URL when printing it to stdout
I don't know why this is required, but somehow requesting the same
string in the browser and in the command line breaks it in the command
line.

This is probably something stupid with the caching on the server
side. If we could figure out how to force the query to run again I
suspect this wouldn't be necessary.