Add a filter option for jobs in various states
Now we can return all the streams and grid cites for a list of jobs
(see previous commit). It's useful to filter for jobs in various
states, to do things like gather the list of inputs to broken jobs, or
to download all the completed job outputs. This adds a --filter option
to do this.